Split up the scripts module into submodules.
